requirements.txt
torch
torchvision
opencv-python
matplotlib
numpy
Pillow
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git